fixes against latest mongoose.
This drops `Model.Query` api because it seems to be internally used by mongoose. Use `Model.query` instead.
